Better modularize.
