+++ index.js
@@ -1,3 +1,2 @@
-// Convenience file to require the SDK from the root of the repository
-module.exports = require('./src/slamscan');
+module.exports = require('./src');
