+++ src/getClamscan.js
@@ -1,0 +1,6 @@
+var config = require('config');
+var Clamscan = require('clamscan');
+
+module.exports = function() {
+  return new Clamscan(config.get('clamscan'));
+};
