+++ src/getS3.js
@@ -1,0 +1,5 @@
+var AWS = require('aws-sdk');
+
+module.exports = function() {
+  return new AWS.S3();
+};
