+++ src/getSns.js
@@ -1,0 +1,5 @@
+var AWS = require('aws-sdk');
+
+module.exports = function() {
+  return new AWS.SNS();
+};
