+++ src/index.js
@@ -1,0 +1,25 @@
+var download = require('./download');
+var downloadClamscanDbFiles = require('./downloadClamscanDbFiles');
+var downloadUrlToFile = require('./downloadUrlToFile');
+var getClamscan = require('./getClamscan');
+var getS3 = require('./getS3');
+var getSns = require('./getSns');
+var handler = require('./handler');
+var initClamscan = require('./initClamscan');
+var manualScan = require('./manualScan');
+var scan = require('./scan');
+var sns = require('./sns');
+
+module.exports = {
+  download: download,
+  downloadClamscanDbFiles: downloadClamscanDbFiles,
+  downloadUrlToFile: downloadUrlToFile,
+  getClamscan: getClamscan,
+  getS3: getS3,
+  getSns: getSns,
+  handler: handler,
+  initClamscan: initClamscan,
+  manualScan: manualScan,
+  scan: scan,
+  sns: sns
+};
